Drop check in while loop
Added a variable for the simulation to only drop the enemy's weapon once

+++ main.py
@@ -5,6 +5,7 @@ hero = Hero(name = "Hero", health = 100)
 hero.equip(iron_sword)
 enemy = Enemy(name = "Enemy", health = 100, weapon = short_bow)
 # The enemy will have a Short Bow as default, although it can be dropped
+enemy_weapon_dropped = False
 
 while hero.health > 0 and enemy.health > 0:
     hero.attack(enemy)
@@ -13,8 +14,9 @@ while hero.health > 0 and enemy.health > 0:
     hero.health_bar.draw()
     enemy.health_bar.draw()
     
-    hero.drop()
-    enemy.drop()
+    if not enemy_weapon_dropped:
+        enemy.drop()
+        enemy_weapon_dropped = True
     input()
     
 if hero.health > enemy.health:
